+++ pages/index.tsx
@@ -1,11 +1,8 @@
 import { GetStaticProps, NextPage } from "next";
 import Head from "next/head";
-import Link from "next/link";
-import { FC } from "react";
 import Parser from "rss-parser";
 import { Paragraph } from "../components/atoms/Paragraph";
-import { SvgPlayArrow } from "../components/atoms/SvgPlayArrow";
-import { useFloatingPlayDispatchContext } from "../context/FloatingPlayAreaContext";
+import { PodcastEpisodeListItem } from "../components/molecules/PodcastEpisodeListItem";
 import { isPodcastEpisodes, PodcastEpisode } from "../types";
 
 type Props = {
@@ -63,56 +60,6 @@ const Page: NextPage<Props> = ({ episodes }) => {
   );
 };
 
-// TODO: あとでコンポーネントにする
-type ListItemProps = {
-  key: string;
-  episode: PodcastEpisode;
-};
-
-const PodcastEpisodeListItem: FC<ListItemProps> = (props) => {
-  const { updateFloatingPlayAreaState } = useFloatingPlayDispatchContext();
-
-  return (
-    <li>
-      <div className="flex gap-4">
-        <div className="flex-shrink-0 relative w-20 h-20 rounded overflow-hidden">
-          <img src="/saru.jpg" width={128} height={128} alt="" />
-
-          <button
-            className="absolute top-0 left-0 w-full h-full flex items-center justify-center bg-gray-700 bg-opacity-20 hover:bg-opacity-5"
-            onClick={() => {
-              updateFloatingPlayAreaState(props.episode);
-            }}
-          >
-            <SvgPlayArrow className="w-10 h-10 fill-current text-gray-50" />
-            <span className="sr-only">再生</span>
-          </button>
-        </div>
-
-        <Link href={`/${props.episode.guid}`}>
-          <a className="flex-grow hover:underline">
-            <div className="grid gap-y-2">
-              <div className="text-xs text-gray-500">
-                {new Date(
-                  Date.parse(props.episode.pubDate)
-                ).toLocaleDateString()}
-              </div>
-
-              <div className="text-base text-gray-900">
-                {props.episode.title}
-              </div>
-
-              <div className="text-xs text-gray-900">
-                {fmtPrettyJPTime(Number(props.episode.itunes.duration))}
-              </div>
-            </div>
-          </a>
-        </Link>
-      </div>
-    </li>
-  );
-};
-
 export default Page;
 
 const parser: Parser = new Parser();
@@ -147,15 +94,3 @@ export const getStaticProps: GetStaticProps<Props> = async () => {
     },
   };
 };
-
-function fmtPrettyJPTime(sec: number): string {
-  const hour = Math.floor(sec / 3600);
-  const min = `${Math.floor(sec / 60) % 60}`.padStart(2, "0");
-  const secOnly = `${sec % 60}`.padStart(2, "0");
-  let ret = "";
-  if (hour > 0) {
-    ret += `${hour}:`;
-  }
-  ret += `${min}:${secOnly}`;
-  return ret;
-}
